move file structure, fix some imports
moved some stuff, fixed some imports, split files into multiple
