+++ .gitignore
@@ -1,2 +1,3 @@
 **/Tuxedo Bake/
 __pycache__
+.vscode
