+++ tools/dictionaries.py
@@ -1,0 +1,111 @@
+# GPL Licence
+
+
+# Bone names from https://github.com/triazo/immersive_scaler/
+# Note from @989onan: Please make sure to make your names are lowercase in this array. I banged my head metaphorically till I figured that out...
+bone_names = {
+    "right_shoulder": ["rightshoulder", "shoulderr", "rshoulder"],
+    "right_arm": ["rightarm", "armr", "rarm", "upperarmr", "rupperarm", "rightupperarm", "uparmr", "ruparm"],
+    "right_elbow": ["rightelbow", "elbowr", "relbow", "lowerarmr", "rightlowerarm", "lowerarmr","rlowerarm", "lowarmr", "rlowarm"],
+    "right_wrist": ["rightwrist", "wristr", "rwrist", "handr", "righthand", "rhand"],
+
+    #hand l fingers
+    "pinkie_0_r": ["littlefinger0r","pinkie0r","rpinkie0","pinkiemetacarpalr"],
+    "pinkie_1_r": ["littlefinger1r","pinkie1r","rpinkie1","pinkieproximalr"],
+    "pinkie_2_r": ["littlefinger2r","pinkie2r","rpinkie2","pinkieintermediater"],
+    "pinkie_3_r": ["littlefinger3r","pinkie3r","rpinkie3","pinkiedistalr"],
+
+    "ring_0_r": ["ringfinger0r","ring0r","rring0","ringmetacarpalr"],
+    "ring_1_r": ["ringfinger1r","ring1r","rring1","ringproximalr"],
+    "ring_2_r": ["ringfinger2r","ring2r","rring2","ringintermediater"],
+    "ring_3_r": ["ringfinger3r","ring3r","rring3","ringdistalr"],
+
+    "middle_0_r": ["middlefinger0r","middle0r","rmiddle0","middlemetacarpalr"],
+    "middle_1_r": ["middlefinger1r","middle1r","rmiddle1","middleproximalr"],
+    "middle_2_r": ["middlefinger2r","middle2r","rmiddle2","middleintermediater"],
+    "middle_3_r": ["middlefinger3r","middle3r","rmiddle3","middledistalr"],
+
+    "index_0_r": ["indexfinger0r","index0r","rindex0","indexmetacarpalr"],
+    "index_1_r": ["indexfinger1r","index1r","rindex1","indexproximalr"],
+    "index_2_r": ["indexfinger2r","index2r","rindex2","indexintermediater"],
+    "index_3_r": ["indexfinger3r","index3r","rindex3","indexdistalr"],
+
+    "thumb_0_r": ["thumb0r","rthumb0","thumbmetacarpalr"],
+    "thumb_1_r": ['thumb0r',"rthumb0","thumbproximalr"],
+    "thumb_2_r": ['thumb1r',"rthumb1","thumbintermediater"],
+    "thumb_3_r": ['thumb2r',"rthumb2","thumbdistalr"],
+
+    "right_leg": ["rightleg", "legr", "rleg", "upperlegr", "rupperleg", "thighr", "rightupperleg", "uplegr", "rupleg"],
+    "right_knee": ["rightknee", "kneer", "rknee", "lowerlegr", "calfr", "rlowerleg", "rcalf", "rightlowerleg", "lowlegr", "rlowleg"],
+    "right_ankle": ["rightankle", "ankler", "rankle", "rightfoot", "footr", "rfoot", "rightfoot", "rightfeet", "feetright", "rfeet", "feetr"],
+    "right_toe": ["righttoe", "toeright", "toer", "rtoe", "toesr", "rtoes"],
+
+    "left_shoulder": ["leftshoulder", "shoulderl", "lshoulder"],
+    "left_arm": ["leftarm", "arml", "rarm", "upperarml", "lupperarm", "leftupperarm", "uparml", "luparm"],
+    "left_elbow": ["leftelbow", "elbowl", "lelbow", "lowerarml", "leftlowerarm", "lowerarml", "llowerarm", "lowarml", "llowarm"],
+    "left_wrist": ["leftwrist", "wristl", "lwrist", "handl", "lefthand", "lhand"],
+
+    #hand l fingers
+
+    "pinkie_0_l": ["pinkiefinger0l","pinkie0l","lpinkie0","pinkiemetacarpall"],
+    "pinkie_1_l": ["littlefinger1l","pinkie1l","lpinkie1","pinkieproximall"],
+    "pinkie_2_l": ["littlefinger2l","pinkie2l","lpinkie2","pinkieintermediatel"],
+    "pinkie_3_l": ["littlefinger3l","pinkie3l","lpinkie3","pinkiedistall"],
+
+    "ring_0_l": ["ringfinger0l","ring0l","lring0","ringmetacarpall"],
+    "ring_1_l": ["ringfinger1l","ring1l","lring1","ringproximall"],
+    "ring_2_l": ["ringfinger2l","ring2l","lring2","ringintermediatel"],
+    "ring_3_l": ["ringfinger3l","ring3l","lring3","ringdistall"],
+
+    "middle_0_l": ["middlefinger0l","middle_0l","lmiddle0","middlemetacarpall"],
+    "middle_1_l": ["middlefinger1l","middle_1l","lmiddle1","middleproximall"],
+    "middle_2_l": ["middlefinger2l","middle_2l","lmiddle2","middleintermediatel"],
+    "middle_3_l": ["middlefinger3l","middle_3l","lmiddle3","middledistall"],
+
+    "index_0_l": ["indexfinger0l","index0l","lindex0","indexmetacarpall"],
+    "index_1_l": ["indexfinger1l","index1l","lindex1","indexproximall"],
+    "index_2_l": ["indexfinger2l","index2l","lindex2","indexintermediatel"],
+    "index_3_l": ["indexfinger3l","index3l","lindex3","indexdistall"],
+
+    "thumb_0_l": ["thumb0l","lthumb0","thumbmetacarpall"],
+    "thumb_1_l": ['thumb0l',"lthumb0","thumbproximall"],
+    "thumb_2_l": ['thumb1l',"lthumb1","thumbintermediatel"],
+    "thumb_3_l": ['thumb2l',"lthumb2","thumbdistall"],
+
+    "left_leg": ["leftleg", "legl", "lleg", "upperlegl", "lupperleg", "thighl", "leftupperleg", "uplegl", "lupleg"],
+    "left_knee": ["leftknee", "kneel", "lknee", "lowerlegl", "llowerleg", "calfl", "lcalf", "leftlowerleg", 'lowlegl', 'llowleg'],
+    "left_ankle": ["leftankle", "anklel", "rankle", "leftfoot", "footl", "lfoot", "leftfoot", "leftfeet", "feetleft", "lfeet", "feetl"],
+    "left_toe": ["lefttoe", "toeleft", "toel", "ltoe", "toesl", "ltoes"],
+
+    "hips": ["pelvis", "hips"],
+    "spine": ["torso", "spine"],
+    "chest": ["chest"],
+    "upper_chest": ["upperchest"],
+    "neck": ["neck"],
+    "head": ["head"],
+    "left_eye": ["eyeleft", "lefteye", "eyel", "leye"],
+    "right_eye": ["eyeright", "righteye", "eyer", "reye"],
+}
+
+# array taken from cats
+dont_delete_these_main_bones = [
+    'Hips', 'Spine', 'Chest', 'Upper Chest', 'Neck', 'Head',
+    'Left leg', 'Left knee', 'Left ankle', 'Left toe',
+    'Right leg', 'Right knee', 'Right ankle', 'Right toe',
+    'Left shoulder', 'Left arm', 'Left elbow', 'Left wrist',
+    'Right shoulder', 'Right arm', 'Right elbow', 'Right wrist',
+    'LeftEye', 'RightEye', 'Eye_L', 'Eye_R',
+    'Left leg 2', 'Right leg 2',
+
+    'Thumb0_L', 'Thumb1_L', 'Thumb2_L',
+    'IndexFinger1_L', 'IndexFinger2_L', 'IndexFinger3_L',
+    'MiddleFinger1_L', 'MiddleFinger2_L', 'MiddleFinger3_L',
+    'RingFinger1_L', 'RingFinger2_L', 'RingFinger3_L',
+    'LittleFinger1_L', 'LittleFinger2_L', 'LittleFinger3_L',
+
+    'Thumb0_R', 'Thumb1_R', 'Thumb2_R',
+    'IndexFinger1_R', 'IndexFinger2_R', 'IndexFinger3_R',
+    'MiddleFinger1_R', 'MiddleFinger2_R', 'MiddleFinger3_R',
+    'RingFinger1_R', 'RingFinger2_R', 'RingFinger3_R',
+    'LittleFinger1_R', 'LittleFinger2_R', 'LittleFinger3_R',
+]
