+++ ui_sections/warnings.py
@@ -4,10 +4,9 @@ import mathutils
 
 from .. import bake as Bake
 from ..properties import get_steam_library
-from ..tools import t, get_meshes_objects, get_armature, simplify_bonename, bone_names
-from ..tools import GenerateTwistBones, TwistTutorialButton, SmartDecimation, RepairShapekeys
-from ..tools import AutoDecimatePresetGood, AutoDecimatePresetQuest, AutoDecimatePresetExcellent
-from ..tools import FitClothes, SRanipal_Labels, has_shapekeys, get_shapekeys_ft, materials_list_update
+from ..tools.translate import t
+from ..tools.core import get_meshes_objects, get_armature, simplify_bonename
+from ..tools.dictionaries import bone_names
 
 from ..ui import register_ui_tab
